Migrate refract tests
